[feat] 지원서 리스트 분리해서 기능 추가
